refactor(iam): split role sync support for complexity gate
